Renaming and applying review comments,
